Moved Daily Reward to minigames
